Move --download test cases to own file
