+++ tests/cases/mod.rs
@@ -1,1 +1,2 @@
+mod download;
 mod logging;
